pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aerosurrogate"
version = "3.0.0"
description = "Aerodynamic surrogate model — FNO + ensemble + digital twin (v3.0)"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    # ML
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    # Optimisation
    "optuna>=3.6.0",
    # Utils
    "requests>=2.31.0",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Boosting (optional — falls back to sklearn GBM)
boosting = [
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
]
# PyTorch (optional — required for FNO field model)
torch = [
    "torch>=2.2.0",
]
# MLflow experiment tracking (optional)
tracking = [
    "mlflow>=2.12.0",
    "pyarrow>=15.0.0",
]
# Kafka digital twin streaming (optional)
streaming = [
    "aiohttp>=3.9.0",
    # confluent-kafka is optional — only needed for Kafka ingest
    # install manually: pip install confluent-kafka
]
# Full API image stack
api = [
    "aerosurrogate[boosting]",
    "aerosurrogate[torch]",
    "aerosurrogate[tracking]",
    "aerosurrogate[streaming]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = [
    "src*", "models*", "data_pipeline*",
    "cfd_pipeline*", "inference*", "optimization*", "digital_twin*",
]

[tool.ruff]
line-length = 100
target-version = "py311"